Allow for new wildfly versions to be extracted
The old code relied on archive's create parameter to know when to run.
Unfortunately it passed the extracted file path to creates and this path
does not change even when you want a new version of wildfly. This mean
that you could not upgrade wildfly using this module without performing
some other act to move the jar aside.

Fix this by not cleaning up the archive file and setting creates to be
the preextracted filename which should be unique per wildfly version.
This has the downside of never cleaning up the archives but using up a
bit of disk space seems preferable to not being able to upgrade
automagically.