module: allow to configure module user/group
Motivation: for security reasons they shouldn't belong to the Wildfly user. Also,
when installing Wildfly via RPM the current implementation fails on `mkdir -p`,
as the `modules` directory belongs to root.

I didn't want to change file ownership handling throughout the whole module. So
this fix allows those who depend on RPMs and those who want to not grant the web
server write access to it's binaries the possibility to achieve that goal.